Fixed events for enabaling and disableing plugins

+++ app/src/plugins/voice/events/ready.ts
@@ -5,9 +5,12 @@ import { dataSource } from "../../../data/datasource";
 import { VoiceChannel } from "../../../data/entities/VoiceChannel";
 import { VoiceActivity } from "../../../data/entities/VoiceActivity";
 
+import { VoicePlugin } from "..";
+
 export const Ready: Event = {
     name: Events.Ready,
     once: true,
+    manualCheck: true,
 
     /**
     * @param {KiwiClient} client
@@ -17,26 +20,34 @@ export const Ready: Event = {
         const VoiceActivityDB = await dataSource.getRepository(VoiceActivity)
 
         var vss = await VoiceChannelsDB.find();
+        console.log(10101010101010101)
 
         for (var vs of vss) {
-            var voiceState = client.guilds.cache.get(vs.guildId)?.voiceStates.cache.get(vs.userId);
-            if (!voiceState) {
-                await VoiceChannelsDB.delete({ userId: vs.userId, guildId: vs.guildId });
+            var voiceState = (await client.guilds.fetch(vs.guildId)).voiceStates.cache.get(vs.userId);
+
+            if (await client.getGuildPlugin(voiceState.guild.id, VoicePlugin.config.name)) {
+                if (!voiceState) {
+                    await VoiceChannelsDB.delete({ userId: vs.userId, guildId: vs.guildId });
+                }
             }
         }
 
         for (var guild of client.guilds.cache.values()) {
-            for (var voiceState of guild.voiceStates.cache.values()) {
+            if (await client.getGuildPlugin(guild.id, VoicePlugin.config.name)) {
+                for (var voiceState of guild.voiceStates.cache.values()) {
 
-                var vs = await VoiceChannelsDB.findOne(
-                    { where: { userId: voiceState.id, guildId: voiceState.guild.id }}
-                );
+                    if (await client.getGuildPlugin(voiceState.guild.id, VoicePlugin.config.name)) {
+                        var vs = await VoiceChannelsDB.findOne(
+                            { where: { userId: voiceState.id, guildId: voiceState.guild.id }}
+                        );
 
-                if (!vs) {
-                    await VoiceChannelsDB.upsert(
-                        { userId: voiceState.id, guildId: voiceState.guild.id, joinTime: new Date() },
-                        ["userId", "guildId"]
-                    );
+                        if (!vs) {
+                            await VoiceChannelsDB.upsert(
+                                { userId: voiceState.id, guildId: voiceState.guild.id, joinTime: new Date() },
+                                ["userId", "guildId"]
+                            );
+                        }
+                    }
                 }
             }
         }
